+++ dataworkspace/dataworkspace/tests/datasets/test_models.py
@@ -1,5 +1,13 @@
-import pytest
+from datetime import datetime
 
+import botocore
+import mock
+import psqlparse
+import psycopg2
+import pytest
+from django.conf import settings
+
+from dataworkspace.apps.core.utils import database_dsn
 from dataworkspace.apps.datasets.models import SourceLink
 from dataworkspace.tests import factories
 
@@ -137,3 +145,148 @@ def test_source_link_filename(db):
         link_type=SourceLink.TYPE_EXTERNAL,
     )
     assert source3.get_filename() == 'a-test-source.csv'
+
+
+def test_dataset_parsed_query_tables(db):
+    ds = factories.DataSetFactory.create(published=True)
+
+    blank_query = factories.CustomDatasetQueryFactory(dataset=ds)
+    assert not blank_query.parsed_query_tables
+
+    standard_query = factories.CustomDatasetQueryFactory(
+        dataset=ds, query='select * from foo'
+    )
+    assert standard_query.parsed_query_tables == ['foo']
+
+    join_query = factories.CustomDatasetQueryFactory(
+        dataset=ds, query='select * from foo join bar on foo.id = bar.id'
+    )
+    assert sorted(join_query.parsed_query_tables) == ['bar', 'foo']
+
+    with_query = factories.CustomDatasetQueryFactory(
+        dataset=ds, query='with test as (select * from foo) select * from test'
+    )
+    assert sorted(with_query.parsed_query_tables) == ['foo', 'test']
+
+    bad_query = factories.CustomDatasetQueryFactory(dataset=ds, query='select * from')
+    with pytest.raises(psqlparse.exceptions.PSqlParseError):
+        bad_query.parsed_query_tables  # pylint: disable=pointless-statement
+
+
+@pytest.fixture
+def metadata_db(db):
+    database = factories.DatabaseFactory(memorable_name='my_database')
+    with psycopg2.connect(
+        database_dsn(settings.DATABASES_DATA['my_database'])
+    ) as conn, conn.cursor() as cursor:
+        cursor.execute(
+            '''
+            CREATE SCHEMA IF NOT EXISTS dataflow;
+            CREATE TABLE IF NOT EXISTS dataflow.metadata (
+                id int, table_schema text, table_name text, dataflow_swapped_tables_utc timestamp
+            );
+            INSERT INTO dataflow.metadata VALUES(1, 'public', 'table1', '2020-09-02 00:01:00.0');
+            INSERT INTO dataflow.metadata VALUES(1, 'public', 'table2', '2020-09-01 00:01:00.0');
+            INSERT INTO dataflow.metadata VALUES(1, 'public', 'table1', '2020-01-01 00:01:00.0');
+            '''
+        )
+        conn.commit()
+        yield database
+        cursor.execute('DROP TABLE dataflow.metadata;')
+
+
+@pytest.mark.django_db
+def test_source_table_data_last_updated(metadata_db):
+    dataset = factories.DataSetFactory()
+    table = factories.SourceTableFactory(
+        dataset=dataset, database=metadata_db, schema='public', table='table1'
+    )
+    assert table.get_data_last_updated_date() == datetime(2020, 9, 2, 0, 1, 0)
+
+    table = factories.SourceTableFactory(
+        dataset=dataset, database=metadata_db, schema='public', table='doesntexist'
+    )
+    assert table.get_data_last_updated_date() is None
+
+
+@pytest.mark.django_db
+def test_source_table_no_metadata_table():
+    table = factories.SourceTableFactory(
+        dataset=factories.DataSetFactory(),
+        database=factories.DatabaseFactory(memorable_name='my_database'),
+        schema='public',
+        table='table1',
+    )
+    assert table.get_data_last_updated_date() is None
+
+
+@pytest.mark.django_db
+def test_custom_query_data_last_updated(metadata_db):
+    dataset = factories.DataSetFactory()
+
+    # Ensure the earliest "last updated" date is returned when
+    # there are multiple tables in the query
+    query = factories.CustomDatasetQueryFactory(
+        dataset=dataset,
+        database=metadata_db,
+        query='select * from table1 join table2 on 1=1',
+    )
+    assert query.get_data_last_updated_date() == datetime(2020, 9, 1, 0, 1, 0)
+
+    # Ensure a single table returns the last update date
+    query = factories.CustomDatasetQueryFactory(
+        dataset=dataset, database=metadata_db, query='select * from table1',
+    )
+    assert query.get_data_last_updated_date() == datetime(2020, 9, 2, 0, 1, 0)
+
+    # Ensure None is returned if we don't have any metadata for the tables
+    query = factories.CustomDatasetQueryFactory(
+        dataset=dataset, database=metadata_db, query='select * from table3',
+    )
+    assert query.get_data_last_updated_date() is None
+
+
+@pytest.mark.django_db
+def test_custom_query_no_metadata_table():
+    query = factories.CustomDatasetQueryFactory(
+        dataset=factories.DataSetFactory(),
+        database=factories.DatabaseFactory(memorable_name='my_database'),
+        query='select * from table1 join table2 on 1=1',
+    )
+    assert query.get_data_last_updated_date() is None
+
+
+@pytest.mark.django_db
+@mock.patch('dataworkspace.apps.datasets.views.boto3.client')
+def test_source_link_data_last_updated(mock_client):
+    dataset = factories.DataSetFactory.create()
+    local_link = factories.SourceLinkFactory(
+        dataset=dataset,
+        link_type=SourceLink.TYPE_LOCAL,
+        url='s3://sourcelink/158776ec-5c40-4c58-ba7c-a3425905ec45/test.txt',
+    )
+
+    # Returns last modified date if the file exists
+    mock_client().head_object.return_value = {
+        'ContentType': 'text/plain',
+        'LastModified': datetime(2020, 9, 2, 0, 1, 0),
+    }
+    assert local_link.get_data_last_updated_date() == datetime(2020, 9, 2, 0, 1, 0)
+
+    # Returns None if file does not exist on s3
+    mock_client().head_object.side_effect = [
+        botocore.exceptions.ClientError(
+            error_response={'Error': {'Message': 'it failed'}},
+            operation_name='head_object',
+        )
+    ]
+    assert local_link.get_data_last_updated_date() is None
+
+    # External links never have a last updated date
+    external_link = factories.SourceLinkFactory(
+        dataset=dataset,
+        link_type=SourceLink.TYPE_EXTERNAL,
+        url='http://www.example.com',
+    )
+    assert external_link.get_data_last_updated_date() is None
+